Adds coutry-code to MapQuest
The MapQuest API returns a two-letter country-code and no country
name. Currently that Country-code is returned as country name when using
the MapQuest-Adapter. But it is not returned as CountryCode though.

This commit solves that by adding the CountryCode.

To handle BC returning that 2 letter code as country-name is kept AS IS.