Synchronize construct with nominatim
Per https://github.com/geocoder-php/nominatim-provider/commit/3068e7f692e0d799679adcfd81df6046cfcab8fb 
Nominatim accepts 2 additional parameters - one being 'required' because of Nominatim Terms of service. Really the constructor should be standard between all providers so if this is accepted for Nominim then it makes sense to put it in the parent. The point of a factory like this is for one syntax to work with many providers

Note I don't think it needs to be 'required'